optimize: some about rpc optimize. (#1935)
