+++ core/src/main/java/io/seata/core/rpc/netty/AbstractRpcRemotingClient.java
@@ -80,7 +80,7 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
                                      ThreadPoolExecutor messageExecutor, NettyPoolKey.TransactionRole transactionRole) {
         super(messageExecutor);
         this.transactionRole = transactionRole;
-        clientBootstrap = new RpcClientBootstrap(nettyClientConfig, eventExecutorGroup, this, transactionRole);
+        clientBootstrap = new RpcClientBootstrap(nettyClientConfig, eventExecutorGroup, transactionRole);
         clientChannelManager = new NettyClientChannelManager(
             new NettyPoolableFactory(this, clientBootstrap), getPoolKeyFunction(), nettyClientConfig);
     }
@@ -105,6 +105,7 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
 
     @Override
     public void init() {
+        clientBootstrap.setChannelHandlers(new ClientHandler());
         clientBootstrap.start();
         timerExecutor.scheduleAtFixedRate(new Runnable() {
             @Override
@@ -129,98 +130,7 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
         if (mergeSendExecutorService != null) {
             mergeSendExecutorService.shutdown();
         }
-    }
-
-    @Override
-    public void channelRead(final ChannelHandlerContext ctx, Object msg) throws Exception {
-        if (!(msg instanceof RpcMessage)) {
-            return;
-        }
-        RpcMessage rpcMessage = (RpcMessage) msg;
-        if (rpcMessage.getBody() == HeartbeatMessage.PONG) {
-            if (LOGGER.isDebugEnabled()) {
-                LOGGER.debug("received PONG from {}", ctx.channel().remoteAddress());
-            }
-            return;
-        }
-        if (rpcMessage.getBody() instanceof MergeResultMessage) {
-            MergeResultMessage results = (MergeResultMessage) rpcMessage.getBody();
-            MergedWarpMessage mergeMessage = (MergedWarpMessage) mergeMsgMap.remove(rpcMessage.getId());
-            for (int i = 0; i < mergeMessage.msgs.size(); i++) {
-                int msgId = mergeMessage.msgIds.get(i);
-                MessageFuture future = futures.remove(msgId);
-                if (future == null) {
-                    if (LOGGER.isInfoEnabled()) {
-                        LOGGER.info("msg: {} is not found in futures.", msgId);
-                    }
-                } else {
-                    future.setResultMessage(results.getMsgs()[i]);
-                }
-            }
-            return;
-        }
-        super.channelRead(ctx, msg);
-    }
-
-    @Override
-    public void dispatch(RpcMessage request, ChannelHandlerContext ctx) {
-        if (clientMessageListener != null) {
-            String remoteAddress = NetUtil.toStringAddress(ctx.channel().remoteAddress());
-            clientMessageListener.onMessage(request, remoteAddress, this);
-        }
-    }
-
-    @Override
-    public void channelInactive(ChannelHandlerContext ctx) throws Exception {
-        if (messageExecutor.isShutdown()) {
-            return;
-        }
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("channel inactive: {}", ctx.channel());
-        }
-        clientChannelManager.releaseChannel(ctx.channel(), NetUtil.toStringAddress(ctx.channel().remoteAddress()));
-        super.channelInactive(ctx);
-    }
-
-    @Override
-    public void userEventTriggered(ChannelHandlerContext ctx, Object evt) {
-        if (evt instanceof IdleStateEvent) {
-            IdleStateEvent idleStateEvent = (IdleStateEvent) evt;
-            if (idleStateEvent.state() == IdleState.READER_IDLE) {
-                if (LOGGER.isInfoEnabled()) {
-                    LOGGER.info("channel {} read idle.", ctx.channel());
-                }
-                try {
-                    String serverAddress = NetUtil.toStringAddress(ctx.channel().remoteAddress());
-                    clientChannelManager.invalidateObject(serverAddress, ctx.channel());
-                } catch (Exception exx) {
-                    LOGGER.error(exx.getMessage());
-                } finally {
-                    clientChannelManager.releaseChannel(ctx.channel(), getAddressFromContext(ctx));
-                }
-            }
-            if (idleStateEvent == IdleStateEvent.WRITER_IDLE_STATE_EVENT) {
-                try {
-                    if (LOGGER.isDebugEnabled()) {
-                        LOGGER.debug("will send ping msg,channel {}", ctx.channel());
-                    }
-                    sendRequest(ctx.channel(), HeartbeatMessage.PING);
-                } catch (Throwable throwable) {
-                    LOGGER.error("send request error: {}", throwable.getMessage(), throwable);
-                }
-            }
-        }
-    }
-
-    @Override
-    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
-        LOGGER.error(FrameworkErrorCode.ExceptionCaught.getErrCode(),
-            NetUtil.toStringAddress(ctx.channel().remoteAddress()) + "connect exception. " + cause.getMessage(), cause);
-        clientChannelManager.releaseChannel(ctx.channel(), getAddressFromChannel(ctx.channel()));
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("remove exception rm channel:{}", ctx.channel());
-        }
-        super.exceptionCaught(ctx, cause);
+        super.destroy();
     }
 
     @Override
@@ -239,12 +149,12 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
     @Override
     public Object sendMsgWithResponse(String serverAddress, Object msg, long timeout)
         throws TimeoutException {
-        return sendAsyncRequestWithResponse(serverAddress, clientChannelManager.acquireChannel(serverAddress), msg, timeout);
+        return super.sendAsyncRequestWithResponse(serverAddress, clientChannelManager.acquireChannel(serverAddress), msg, timeout);
     }
 
     @Override
     public void sendResponse(RpcMessage request, String serverAddress, Object msg) {
-        super.sendResponse(request, clientChannelManager.acquireChannel(serverAddress), msg);
+        super.defaultSendResponse(request, clientChannelManager.acquireChannel(serverAddress), msg);
     }
 
     /**
@@ -321,7 +231,7 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
                     Channel sendChannel = null;
                     try {
                         sendChannel = clientChannelManager.acquireChannel(address);
-                        sendRequest(sendChannel, mergeMessage);
+                        AbstractRpcRemotingClient.super.defaultSendRequest(sendChannel, mergeMessage);
                     } catch (FrameworkException e) {
                         if (e.getErrcode() == FrameworkErrorCode.ChannelIsNotWritable && sendChannel != null) {
                             destroyChannel(address, sendChannel);
@@ -358,4 +268,103 @@ public abstract class AbstractRpcRemotingClient extends AbstractRpcRemoting
             }
         }
     }
+
+    /**
+     * The type ClientHandler.
+     */
+    class ClientHandler extends AbstractHandler {
+
+        @Override
+        public void dispatch(RpcMessage request, ChannelHandlerContext ctx) {
+            if (clientMessageListener != null) {
+                String remoteAddress = NetUtil.toStringAddress(ctx.channel().remoteAddress());
+                clientMessageListener.onMessage(request, remoteAddress);
+            }
+        }
+
+        @Override
+        public void channelRead(final ChannelHandlerContext ctx, Object msg) throws Exception {
+            if (!(msg instanceof RpcMessage)) {
+                return;
+            }
+            RpcMessage rpcMessage = (RpcMessage) msg;
+            if (rpcMessage.getBody() == HeartbeatMessage.PONG) {
+                if (LOGGER.isDebugEnabled()) {
+                    LOGGER.debug("received PONG from {}", ctx.channel().remoteAddress());
+                }
+                return;
+            }
+            if (rpcMessage.getBody() instanceof MergeResultMessage) {
+                MergeResultMessage results = (MergeResultMessage) rpcMessage.getBody();
+                MergedWarpMessage mergeMessage = (MergedWarpMessage) mergeMsgMap.remove(rpcMessage.getId());
+                for (int i = 0; i < mergeMessage.msgs.size(); i++) {
+                    int msgId = mergeMessage.msgIds.get(i);
+                    MessageFuture future = futures.remove(msgId);
+                    if (future == null) {
+                        if (LOGGER.isInfoEnabled()) {
+                            LOGGER.info("msg: {} is not found in futures.", msgId);
+                        }
+                    } else {
+                        future.setResultMessage(results.getMsgs()[i]);
+                    }
+                }
+                return;
+            }
+            super.channelRead(ctx, msg);
+        }
+
+        @Override
+        public void channelInactive(ChannelHandlerContext ctx) throws Exception {
+            if (messageExecutor.isShutdown()) {
+                return;
+            }
+            if (LOGGER.isInfoEnabled()) {
+                LOGGER.info("channel inactive: {}", ctx.channel());
+            }
+            clientChannelManager.releaseChannel(ctx.channel(), NetUtil.toStringAddress(ctx.channel().remoteAddress()));
+            super.channelInactive(ctx);
+        }
+
+        @Override
+        public void userEventTriggered(ChannelHandlerContext ctx, Object evt) {
+            if (evt instanceof IdleStateEvent) {
+                IdleStateEvent idleStateEvent = (IdleStateEvent)evt;
+                if (idleStateEvent.state() == IdleState.READER_IDLE) {
+                    if (LOGGER.isInfoEnabled()) {
+                        LOGGER.info("channel {} read idle.", ctx.channel());
+                    }
+                    try {
+                        String serverAddress = NetUtil.toStringAddress(ctx.channel().remoteAddress());
+                        clientChannelManager.invalidateObject(serverAddress, ctx.channel());
+                    } catch (Exception exx) {
+                        LOGGER.error(exx.getMessage());
+                    } finally {
+                        clientChannelManager.releaseChannel(ctx.channel(), getAddressFromContext(ctx));
+                    }
+                }
+                if (idleStateEvent == IdleStateEvent.WRITER_IDLE_STATE_EVENT) {
+                    try {
+                        if (LOGGER.isDebugEnabled()) {
+                            LOGGER.debug("will send ping msg,channel {}", ctx.channel());
+                        }
+                        AbstractRpcRemotingClient.super.defaultSendRequest(ctx.channel(), HeartbeatMessage.PING);
+                    } catch (Throwable throwable) {
+                        LOGGER.error("send request error: {}", throwable.getMessage(), throwable);
+                    }
+                }
+            }
+        }
+
+        @Override
+        public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
+            LOGGER.error(FrameworkErrorCode.ExceptionCaught.getErrCode(),
+                NetUtil.toStringAddress(ctx.channel().remoteAddress()) + "connect exception. " + cause.getMessage(), cause);
+            clientChannelManager.releaseChannel(ctx.channel(), getAddressFromChannel(ctx.channel()));
+            if (LOGGER.isInfoEnabled()) {
+                LOGGER.info("remove exception rm channel:{}", ctx.channel());
+            }
+            super.exceptionCaught(ctx, cause);
+        }
+    }
+
 }
